--- Projects/Common/Sources/Extensions/Bundle+Version.swift
@@ -1,14 +1,0 @@
-//
-//  Bundle+Version.swift
-//  CommonManifests
-//
-//  Created by Kim SungHun on 2023/09/24.
-//
-
-import Foundation
-
-extension Bundle {
-	var version: String {
-		return ""
-	}
-}
