+++ Projects/Core/Sources/Common/Extensions/Bundle+.swift
@@ -1,0 +1,15 @@
+//
+//  Bundle+.swift
+//  Core
+//
+//  Created by Kim SungHun on 2023/10/23.
+//  Copyright © 2023 com.tenten. All rights reserved.
+//
+
+import Foundation
+
+extension Bundle {
+	var version: String {
+		return ""
+	}
+}
